+++ app/src/main/java/me/partypronl/quibble/data/models/JournalEntryModel.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "journal_entry")
 data class JournalEntryModel(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val type: String,
     val date: Long
 )
